# feat: multithread linting