--- a/lib/cli.js
+++ b/lib/cli.js
@@ -67,7 +67,7 @@ function countErrors(results) {
 /**
  * Creates an option module from the provided CLI options and encodes it as a data URL.
  * @param {ParsedCLIOptions} options The CLI options.
- * @returns {string} The URL of the option module.
+ * @returns {URL} The URL of the option module.
  */
 function createOptionModule(options) {
 	const translateOptionsFileURL = new URL(
@@ -79,7 +79,9 @@ function createOptionModule(options) {
 		`export default await translateOptions(${JSON.stringify(options)});\n`;
 
 	// Base64 encoding is typically shorter than URL encoding
-	return `data:text/javascript;base64,${Buffer.from(optionsSrc).toString("base64")}`;
+	return new URL(
+		`data:text/javascript;base64,${Buffer.from(optionsSrc).toString("base64")}`,
+	);
 }
 
 /**