--- a/lib/cli.js
+++ b/lib/cli.js
@@ -16,9 +16,9 @@
 //------------------------------------------------------------------------------
 
 const fs = require("node:fs"),
+	{ mkdir, stat, writeFile } = require("node:fs/promises"),
 	path = require("node:path"),
 	{ pathToFileURL } = require("node:url"),
-	{ promisify } = require("node:util"),
 	{ LegacyESLint } = require("./eslint"),
 	{
 		ESLint,
@@ -45,10 +45,6 @@ const debug = require("debug")("eslint:cli");
 // Helpers
 //------------------------------------------------------------------------------
 
-const mkdir = promisify(fs.mkdir);
-const stat = promisify(fs.stat);
-const writeFile = promisify(fs.writeFile);
-
 /**
  * Count error messages.
  * @param {LintResult[]} results The lint results.
@@ -66,6 +62,24 @@ function countErrors(results) {
 	}
 
 	return { errorCount, fatalErrorCount, warningCount };
+}
+
+/**
+ * Creates an option module from the provided CLI options and encodes it as a data URL.
+ * @param {ParsedCLIOptions} options The CLI options.
+ * @returns {string} The URL of the option module.
+ */
+function createOptionModule(options) {
+	const translateOptionsFileURL = new URL(
+		"./shared/translate-cli-options.js",
+		pathToFileURL(__filename),
+	).href;
+	const optionsSrc =
+		`import translateOptions from ${JSON.stringify(translateOptionsFileURL)};\n` +
+		`export default await translateOptions(${JSON.stringify(options)});\n`;
+
+	// Base64 encoding is typically shorter than URL encoding
+	return `data:text/javascript;base64,${Buffer.from(optionsSrc).toString("base64")}`;
 }
 
 /**
@@ -390,17 +404,7 @@ const cli = {
 		let engine;
 
 		if (options.concurrency && options.concurrency !== "off") {
-			const fileURL = new URL(
-				"./shared/translate-cli-options.js",
-				pathToFileURL(__filename),
-			).href;
-			const optionsSrc =
-				`import translateOptions from ${JSON.stringify(fileURL)};\n` +
-				`export default await translateOptions(${JSON.stringify(options)});\n`;
-
-			// Base64 encoding is typically shorter than URL encoding
-			const optionsURL = `data:text/javascript;base64,${Buffer.from(optionsSrc).toString("base64")}`;
-
+			const optionsURL = createOptionModule(options);
 			engine = await ESLint.fromOptionModule(optionsURL);
 		} else {
 			const eslintOptions = await translateOptions(