--- a/lib/cli.js
+++ b/lib/cli.js
@@ -65,11 +65,11 @@ function countErrors(results) {
 }
 
 /**
- * Creates an option module from the provided CLI options and encodes it as a data URL.
+ * Creates an options module from the provided CLI options and encodes it as a data URL.
  * @param {ParsedCLIOptions} options The CLI options.
- * @returns {URL} The URL of the option module.
+ * @returns {URL} The URL of the options module.
  */
-function createOptionModule(options) {
+function createOptionsModule(options) {
 	const translateOptionsFileURL = new URL(
 		"./shared/translate-cli-options.js",
 		pathToFileURL(__filename),
@@ -406,8 +406,8 @@ const cli = {
 		let engine;
 
 		if (options.concurrency && options.concurrency !== "off") {
-			const optionsURL = createOptionModule(options);
-			engine = await ESLint.fromOptionModule(optionsURL);
+			const optionsURL = createOptionsModule(options);
+			engine = await ESLint.fromOptionsModule(optionsURL);
 		} else {
 			const eslintOptions = await translateOptions(
 				options,