--- a/lib/cli.js
+++ b/lib/cli.js
@@ -76,7 +76,7 @@ function createOptionsModule(options) {
 	).href;
 	const optionsSrc =
 		`import translateOptions from ${JSON.stringify(translateOptionsFileURL)};\n` +
-		`export default await translateOptions(${JSON.stringify(options)});\n`;
+		`export default await translateOptions(${JSON.stringify(options)}, "flat");\n`;
 
 	// Base64 encoding is typically shorter than URL encoding
 	return new URL(
@@ -283,8 +283,8 @@ const cli = {
 			}
 
 			const engine = usingFlatConfig
-				? new ESLint(await translateOptions(options))
-				: new LegacyESLint(await translateOptions(options, false));
+				? new ESLint(await translateOptions(options, "flat"))
+				: new LegacyESLint(await translateOptions(options));
 			const fileConfig = await engine.calculateConfigForFile(
 				options.printConfig,
 			);
@@ -299,7 +299,7 @@ const cli = {
 			);
 
 			try {
-				const flatOptions = await translateOptions(options);
+				const flatOptions = await translateOptions(options, "flat");
 				const spawn = require("cross-spawn");
 				const flags = await cli.calculateInspectConfigFlags(
 					flatOptions.overrideConfigFile,
@@ -411,7 +411,7 @@ const cli = {
 		} else {
 			const eslintOptions = await translateOptions(
 				options,
-				usingFlatConfig,
+				usingFlatConfig ? "flat" : "eslintrc",
 			);
 			engine = new ActiveESLint(eslintOptions);
 		}