--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -11,6 +11,7 @@
 
 const path = require("node:path");
 const fs = require("node:fs");
+const { isMainThread, threadId } = require("node:worker_threads");
 const fsp = fs.promises;
 const isGlob = require("is-glob");
 const hash = require("../cli-engine/hash");
@@ -21,6 +22,7 @@ const { Linter } = require("../linter");
 const { getShorthandName } = require("../shared/naming");
 const LintResultCache = require("../cli-engine/lint-result-cache");
 const { ConfigLoader, LegacyConfigLoader } = require("../config/config-loader");
+const createDebug = require("debug");
 
 //-----------------------------------------------------------------------------
 // Fixup references
@@ -56,7 +58,13 @@ const MINIMATCH_OPTIONS = { dot: true };
 // Helpers
 //------------------------------------------------------------------------------
 
-const debug = require("debug")("eslint:eslint-helpers");
+const hrtimeBigint = process.hrtime.bigint;
+
+createDebug.formatters.t = timeDiff =>
+	`${(timeDiff + 500_000n) / 1_000_000n} ms`;
+const debug = createDebug(
+	`eslint:eslint-helpers${isMainThread ? "" : `:thread-${threadId}`}`,
+);
 
 //-----------------------------------------------------------------------------
 // Errors
@@ -893,7 +901,9 @@ function processOptions({
 		concurrency !== "auto" &&
 		!isPositiveInteger(concurrency)
 	) {
-		errors.push("'concurrency' must be positive integer, 'auto' or 'off'.");
+		errors.push(
+			'\'concurrency\' must be a positive integer, "auto", or "off".',
+		);
 	}
 	if (!isNonEmptyString(cwd) || !path.isAbsolute(cwd)) {
 		errors.push("'cwd' must be an absolute path.");
@@ -1150,9 +1160,9 @@ function verifyText({
 	stats,
 	linter,
 }) {
-	const filePath = providedFilePath || "<text>";
+	const startTime = hrtimeBigint();
 
-	debug(`Lint ${filePath}`);
+	const filePath = providedFilePath || "<text>";
 
 	/*
 	 * Verify.
@@ -1203,6 +1213,9 @@ function verifyText({
 			fixPasses: linter.getFixPassCount(),
 		};
 	}
+
+	const endTime = hrtimeBigint();
+	debug('File "%s" linted in %t', filePath, endTime - startTime);
 
 	return result;
 }
@@ -1279,10 +1292,17 @@ async function lintFile(
 	return retrier
 		.retry(
 			async () => {
+				const beforeReadFile = hrtimeBigint();
 				const text = await fsp.readFile(filePath, {
 					encoding: "utf8",
 					signal: controller?.signal,
 				});
+				const afterReadFile = hrtimeBigint();
+				debug(
+					'File "%s" read in %t',
+					filePath,
+					afterReadFile - beforeReadFile,
+				);
 
 				// fail immediately if an error occurred in another file
 				controller?.signal.throwIfAborted();