--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -1008,7 +1008,7 @@ function processOptions({
 /**
  * Loads ESLint constructor options from an option module.
  * @param {string} optionsURL The URL string of the option module to load.
- * @returns {ESLintOptions} ESLint constructor options.
+ * @returns {Promise<ESLintOptions>} ESLint constructor options.
  */
 async function loadOptionsFromModule(optionsURL) {
 	return (await import(optionsURL)).default;