--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -1007,7 +1007,7 @@ function processOptions({
 
 /**
  * Loads ESLint constructor options from an option module.
- * @param {string|URL} optionsURL The URL of the option module to load.
+ * @param {string} optionsURL The URL string of the option module to load.
  * @returns {ESLintOptions} ESLint constructor options.
  */
 async function loadOptionsFromModule(optionsURL) {