--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -17,7 +17,6 @@ const isGlob = require("is-glob");
 const hash = require("../cli-engine/hash");
 const minimatch = require("minimatch");
 const globParent = require("glob-parent");
-const { Retrier } = require("@humanwhocodes/retry");
 const { Linter } = require("../linter");
 const { getShorthandName } = require("../shared/naming");
 const LintResultCache = require("../cli-engine/lint-result-cache");
@@ -40,6 +39,7 @@ const MINIMATCH_OPTIONS = { dot: true };
  * @import { Config as CalculatedConfig } from "../config/config.js";
  * @import { FlatConfigArray } from "../config/flat-config-array.js";
  * @import { WarningService } from "../services/warning-service.js";
+ * @import { Retrier } from "@humanwhocodes/retry";
  */
 
 /** @typedef {import("../types").Linter.Config} Config */
@@ -1227,7 +1227,7 @@ function verifyText({
  * @param {ESLintOptions} eslintOptions The processed ESLint options.
  * @param {Linter} linter The linter instance to use.
  * @param {?LintResultCache} lintResultCache The result cache or `null`.
- * @param {Retrier} retrier Used to retry linting on certain errors.
+ * @param {Retrier} [retrier] Used to retry linting on certain errors.
  * @param {AbortController} [controller] Used to stop linting when an error occurs.
  * @returns {Promise<LintResult>} The lint result.
  */
@@ -1289,43 +1289,45 @@ async function lintFile(
 	// set up fixer for fixTypes if necessary
 	const fixer = getFixerForFixTypes(fix, fixTypesSet, config);
 
-	return retrier
-		.retry(
-			async () => {
-				const beforeReadFile = hrtimeBigint();
-				const text = await fsp.readFile(filePath, {
-					encoding: "utf8",
-					signal: controller?.signal,
-				});
-				const afterReadFile = hrtimeBigint();
-				debug(
-					'File "%s" read in %t',
-					filePath,
-					afterReadFile - beforeReadFile,
-				);
-
-				// fail immediately if an error occurred in another file
-				controller?.signal.throwIfAborted();
-
-				// do the linting
-				return verifyText({
-					text,
-					filePath,
-					configs,
-					cwd,
-					fix: fixer,
-					allowInlineConfig,
-					ruleFilter,
-					stats,
-					linter,
-				});
-			},
-			{ signal: controller?.signal },
-		)
-		.catch(error => {
-			controller?.abort(error);
-			throw error;
+	/**
+	 * Reads the file and lints its content.
+	 * @returns {Promise<LintResult>} A lint result.
+	 */
+	async function readAndVerifyFile() {
+		const beforeReadFile = hrtimeBigint();
+		const text = await fsp.readFile(filePath, {
+			encoding: "utf8",
+			signal: controller?.signal,
 		});
+		const afterReadFile = hrtimeBigint();
+		debug('File "%s" read in %t', filePath, afterReadFile - beforeReadFile);
+
+		// fail immediately if an error occurred in another file
+		controller?.signal.throwIfAborted();
+
+		// do the linting
+		return verifyText({
+			text,
+			filePath,
+			configs,
+			cwd,
+			fix: fixer,
+			allowInlineConfig,
+			ruleFilter,
+			stats,
+			linter,
+		});
+	}
+
+	// Use the retrier if provided, otherwise just call the function.
+	const readAndVerifyFilePromise = retrier
+		? retrier.retry(readAndVerifyFile, { signal: controller?.signal })
+		: readAndVerifyFile();
+
+	return readAndVerifyFilePromise.catch(error => {
+		controller?.abort(error);
+		throw error;
+	});
 }
 
 /**
@@ -1419,18 +1421,6 @@ function createConfigLoader(
 		: new LegacyConfigLoader(configLoaderOptions);
 }
 
-const lintFileRetryCodes = new Set(["ENFILE", "EMFILE"]);
-
-/**
- * Creates a retrier to use with `lintFile`.
- * @param {Object} [options] The options for the retrier.
- * @param {number} [options.concurrency] The maximum number of concurrent tasks.
- * @returns {Retrier} A new retrier instance configured to retry on ENFILE and EMFILE errors.
- */
-function createLintFileRetrier(options) {
-	return new Retrier(error => lintFileRetryCodes.has(error.code), options);
-}
-
 //-----------------------------------------------------------------------------
 // Exports
 //-----------------------------------------------------------------------------
@@ -1456,7 +1446,6 @@ module.exports = {
 	verifyText,
 	lintFile,
 	createLinter,
-	createLintFileRetrier,
 	createDefaultConfigs,
 	createConfigLoader,
 };