--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -569,6 +569,7 @@ async function findFiles({
 		filePaths.map(filePath => fsp.stat(filePath).catch(() => {})),
 	);
 
+	const promises = [];
 	stats.forEach((stat, index) => {
 		const filePath = filePaths[index];
 		const pattern = normalizeToPosix(patterns[index]);
@@ -577,6 +578,7 @@ async function findFiles({
 			// files are added directly to the list
 			if (stat.isFile()) {
 				results.push(filePath);
+				promises.push(configLoader.loadConfigArrayForFile(filePath));
 			}
 
 			// directories need extensions attached
@@ -622,11 +624,14 @@ async function findFiles({
 	}
 
 	// now we are safe to do the search
-	const globbyResults = await globMultiSearch({
-		searches,
-		configLoader,
-		errorOnUnmatchedPattern,
-	});
+	promises.push(
+		globMultiSearch({
+			searches,
+			configLoader,
+			errorOnUnmatchedPattern,
+		}),
+	);
+	const globbyResults = (await Promise.all(promises)).at(-1);
 
 	return [...new Set([...results, ...globbyResults])];
 }