--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -1232,6 +1232,7 @@ function verifyText({
  * @param {ESLintOptions} eslintOptions The processed ESLint options.
  * @param {Linter} linter The linter instance to use.
  * @param {?LintResultCache} lintResultCache The result cache or `null`.
+ * @param {?{ duration: bigint; }} readFileCounter Used to keep track of the time spent reading files.
  * @param {Retrier} [retrier] Used to retry linting on certain errors.
  * @param {AbortController} [controller] Used to stop linting when an error occurs.
  * @returns {Promise<LintResult>} The lint result.
@@ -1242,6 +1243,7 @@ async function lintFile(
 	eslintOptions,
 	linter,
 	lintResultCache,
+	readFileCounter,
 	retrier,
 	controller,
 ) {
@@ -1305,11 +1307,11 @@ async function lintFile(
 			signal: controller?.signal,
 		});
 		const readFileExitTime = hrtimeBigint();
-		debug(
-			'File "%s" read in %t',
-			filePath,
-			readFileExitTime - readFileEnterTime,
-		);
+		const readFileDuration = readFileExitTime - readFileEnterTime;
+		debug('File "%s" read in %t', filePath, readFileDuration);
+		if (readFileCounter) {
+			readFileCounter.duration += readFileDuration;
+		}
 
 		// fail immediately if an error occurred in another file
 		controller?.signal.throwIfAborted();