--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -1299,13 +1299,17 @@ async function lintFile(
 	 * @returns {Promise<LintResult>} A lint result.
 	 */
 	async function readAndVerifyFile() {
-		const beforeReadFile = hrtimeBigint();
+		const readFileEnterTime = hrtimeBigint();
 		const text = await fsp.readFile(filePath, {
 			encoding: "utf8",
 			signal: controller?.signal,
 		});
-		const afterReadFile = hrtimeBigint();
-		debug('File "%s" read in %t', filePath, afterReadFile - beforeReadFile);
+		const readFileExitTime = hrtimeBigint();
+		debug(
+			'File "%s" read in %t',
+			filePath,
+			readFileExitTime - readFileEnterTime,
+		);
 
 		// fail immediately if an error occurred in another file
 		controller?.signal.throwIfAborted();