--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -55,7 +55,7 @@ const MINIMATCH_OPTIONS = { dot: true };
  */
 
 //------------------------------------------------------------------------------
-// Helpers
+// Internal Helpers
 //------------------------------------------------------------------------------
 
 const hrtimeBigint = process.hrtime.bigint;