--- a/lib/eslint/eslint-helpers.js
+++ b/lib/eslint/eslint-helpers.js
@@ -1006,8 +1006,8 @@ function processOptions({
 }
 
 /**
- * Loads ESLint constructor options from an option module.
- * @param {string} optionsURL The URL string of the option module to load.
+ * Loads ESLint constructor options from an options module.
+ * @param {string} optionsURL The URL string of the options module to load.
  * @returns {Promise<ESLintOptions>} ESLint constructor options.
  */
 async function loadOptionsFromModule(optionsURL) {