--- a/lib/eslint/eslint.js
+++ b/lib/eslint/eslint.js
@@ -885,6 +885,8 @@ class ESLint {
 
 		/** @type {LintResult[]} */
 		let results;
+
+		// The value of `module.exports.calculateWorkerCount` can be overridden in tests.
 		const workerCount = module.exports.calculateWorkerCount(
 			concurrency,
 			filePaths.length,