--- a/tests/fixtures/option-modules/empty.mjs
+++ b/tests/fixtures/option-modules/empty.mjs
@@ -1,0 +1,1 @@
+export default {};