--- a/tests/fixtures/option-modules/test-only-flag.cjs
+++ b/tests/fixtures/option-modules/test-only-flag.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+    flags: ["test_only"],
+};