… .github/workflows/types-integration.yml
… docs/src/integrate/nodejs-api.md
… docs/src/use/command-line-interface.md
… lib/eslint/worker.js
… lib/options.js
… lib/services/warning-service.js
… lib/shared/translate-cli-options.js
… lib/types/index.d.ts
… tests/bin/eslint.js
… tests/fixtures/option-modules/test-only-enum-flag.cts
… tests/fixtures/option-modules/test-only-enum-flag.mts
… tests/fixtures/option-modules/test-only-flag.cts
… tests/fixtures/option-modules/test-only-flag.mjs
… tests/fixtures/option-modules/test-only-flag.mts
… tests/lib/cli.js
… tests/lib/eslint/eslint.js
… tests/lib/services/warning-service.js
… tests/lib/types/types.test.ts