<!--
    Thank you for contributing!

    ESLint adheres to the [OpenJS Foundation Code of Conduct](https://eslint.org/conduct).
-->

#### Prerequisites checklist

- [X] I have read the [contributing guidelines](https://github.com/eslint/eslint/blob/HEAD/CONTRIBUTING.md).

#### What is the purpose of this pull request? (put an "X" next to an item)

<!--
    The following template is intentionally not a markdown checkbox list for the reasons
    explained in https://github.com/eslint/eslint/pull/12848#issuecomment-580302888
-->

[ ] Documentation update
[ ] Bug fix ([template](https://raw.githubusercontent.com/eslint/eslint/HEAD/templates/bug-report.md))
[ ] New rule ([template](https://raw.githubusercontent.com/eslint/eslint/HEAD/templates/rule-proposal.md))
[ ] Changes an existing rule ([template](https://raw.githubusercontent.com/eslint/eslint/HEAD/templates/rule-change-proposal.md))
[ ] Add autofix to a rule
[ ] Add a CLI option
[X] Add something to the core
[ ] Other, please explain:

<!--
    If the item you've checked above has a template, please paste the template questions below and answer them. (If this pull request is addressing an issue, you can just paste a link to the issue here instead.)
-->

closes #3565

<!--
    Please ensure your pull request is ready:

    - Read the pull request guide (https://eslint.org/docs/latest/contribute/pull-requests)
    - Include tests for this change
    - Update documentation for this change (if appropriate)
-->

<!--
    The following is required for all pull requests:
-->

#### What changes did you make? (Give an overview)

This pull request implements eslint/rfcs#129, featuring multithread linting and option modules.

#### Is there anything you'd like reviewers to focus on?

---

#### Testing the prototype

To test this in a local package:

```shell
npm i -D eslint/eslint#rfc129
node ./node_modules/.bin/eslint --concurrency=auto
```

#### Debug Messages

These are some interesting debug messages.

Number of enumerated files, time spent to enumerate files in the main thread, and effective number of worker threads:

```text
2025-06-06T06:12:17.489Z eslint:eslint 861 files found in: 1861ms
2025-06-06T06:12:17.489Z eslint:eslint Linting using 4 worker thread(s).
```

Alternatively, if no concurrency is used:

```text
2025-06-06T06:30:37.419Z eslint:eslint Linting without concurrency.
```

Time to load dependencies in a worker thread:

```text
2025-06-06T06:20:33.394Z eslint:worker:thread-1 Dependencies loaded in 206 ms
2025-06-06T06:20:33.400Z eslint:worker:thread-4 Dependencies loaded in 214 ms
2025-06-06T06:20:33.400Z eslint:worker:thread-2 Dependencies loaded in 211 ms
2025-06-06T06:20:33.407Z eslint:worker:thread-3 Dependencies loaded in 217 ms
```

Additional time to initialize a worker thread after dependencies are loaded:

```text
2025-06-06T06:20:33.407Z eslint:worker:thread-1 Linting started 13 ms after dependencies loaded
2025-06-06T06:20:33.415Z eslint:worker:thread-4 Linting started 15 ms after dependencies loaded
2025-06-06T06:20:33.417Z eslint:worker:thread-2 Linting started 18 ms after dependencies loaded
2025-06-06T06:20:33.425Z eslint:worker:thread-3 Linting started 18 ms after dependencies loaded
```

Time to load a config array in a worker thread. It takes long to load a config array the first time, successive loads use the cache and are much faster:

```text
2025-06-06T06:20:35.182Z eslint:worker:thread-3 Config array for file "...\eslint\.github\ISSUE_TEMPLATE\bug-report.yml" loaded in 1757 ms
2025-06-06T06:20:35.186Z eslint:worker:thread-4 Config array for file "...\eslint\.codeclimate.yml" loaded in 1771 ms
2025-06-06T06:20:35.194Z eslint:worker:thread-1 Config array for file "...\eslint\.c8rc" loaded in 1788 ms
```

Time to read a file to be linted:

```text
2025-06-06T06:20:35.208Z eslint:eslint-helpers:thread-3 File "...\eslint\.github\ISSUE_TEMPLATE\bug-report.yml" read in 1 ms
2025-06-06T06:20:35.209Z eslint:eslint-helpers:thread-4 File "...\eslint\.codeclimate.yml" read in 1 ms
2025-06-06T06:20:35.211Z eslint:eslint-helpers:thread-1 File "...\eslint\.c8rc" read in 1 ms
```

Time spent by the linter on a file:

```text
2025-06-06T06:20:35.218Z eslint:eslint-helpers:thread-1 File "...\eslint\.c8rc" linted in 7 ms
```

Total time spent on a file in a worker thread. This includes loading the config array, reading the file, linting the file, and a little overhead:

```text
2025-06-06T06:20:35.218Z eslint:worker:thread-1 File "...\eslint\.c8rc" processed in 1811 ms
```

<a name="benchmark-references"></a>
#### Benchmark References

**ESLint**
Repo: https://github.com/eslint/eslint
Files: &lt;1000
Commands:
```shell
git clone --no-tags --single-branch https://github.com/eslint/eslint -b rfc129
cd eslint
npm i
npm --prefix docs i
node bin/eslint.js --concurrency=auto
```

**OpenUI5**
Repo: https://github.com/mauriciolauffer/openui5/tree/eslint-v9
Files: ~12500
Commands:
```shell
git clone --no-tags --single-branch https://github.com/mauriciolauffer/openui5 -b eslint-v9
cd openui5
npm i -D eslint/eslint#rfc129
node node_modules/eslint/bin/eslint.js --concurrency=auto src --quiet
```

**core-js**
Repo: https://github.com/zloirock/core-js
Files: ~8000
Commands:
```shell
git clone --no-tags --single-branch https://github.com/zloirock/core-js
cd core-js
npm i
npm --prefix tests/eslint i -D eslint/eslint#rfc129
# Fails on Windows
node tests/eslint/node_modules/eslint/bin/eslint.js --concurrency=auto --config=tests/eslint/eslint.config.js
```

#### Archived Topics

<details>
<a name="eslintgetrulesmetaforresults"></a>
#### `ESLint#getRulesMetaForResults()`

This section describes three different solutions, each implemented in its own branch, for the handling of `ESLint#getRulesMetaForResults()` as discussed in the RFC (see [`ESLint#getRulesMetaForResults()`](https://github.com/eslint/rfcs/tree/main/designs/2024-multithread-linting#eslintgetrulesmetaforresults)).

To recap, the `findFiles` function currently loads config files for files it finds by resolving glob patterns, and it will continue to do so in multithread mode. But if a file path is passed explicitly (i.e. not matched by a glob pattern), then `findFiles` does not try to load a config file for it. The problem with the `getRulesMetaForResults`  method in its present implementation, is that this method requires config files for all linted files to have been loaded in the current thread. This means that in order for `getRulesMetaForResults` to continue working in multithread mode, either config files for all explicitly passed files must be loaded in the controlling thread — additionally to worker threads — during the `lintFiles` invocation, or `getRulesMetaForResults` must be changed to have the information it requires calculated by worker threads.

**[rfc129 (this pull request)](https://github.com/eslint/eslint/tree/rfc129)** 
This is based on [Solution 1](https://github.com/eslint/rfcs/tree/main/designs/2024-multithread-linting#solution-1-preload-config-arrays-in-the-controlling-thread) shown in the RFC. This implementation ensures that the config files for all enumerated files are loaded in the controlling thread while worker threads are running. If `getRulesMetaForResults` is not called after `lintFiles`, it's possible that some config files will be loaded unnecessarily.

**[rfc129-b](https://github.com/eslint/eslint/tree/rfc129-b)**
Based on https://github.com/eslint/rfcs/pull/129#discussion_r2092880643. All config files are loaded by `findFiles` before the worker threads are created, without using limited concurrency. This is similar to the previous implementation in terms of performance, but it simplifies the code and ensures that later access to config files can be done synchronously.

**[rfc129-c](https://github.com/eslint/eslint/tree/rfc129-c)**
Based on [Solution 2](https://github.com/eslint/rfcs/tree/main/designs/2024-multithread-linting#solution-2-merge-precalculated-rules-meta-in-the-controlling-thread) shown in the RFC.
Rules `meta` objects and `usedDeprecatedRules` for all results are precalculated in the worker threads. This solution requires some important refactiongs and it adds complexity to the existing logic, for example an additional field on the `ESLint` instance to store rules `meta` objects or a `BroadcastChannel` to send warnings to the main thread. The advantage is that config files for explicitly passed filenames don't need to be loaded in the controlling thread. This yields a measurable performance improvement especially when filenames are passed explicitly to `lintFiles`.

<a name="design-questions"></a>
#### Design Questions

<s>**Handling `--concurrency=1`**
Running ESLint from the CLI with `--concurrency=1` performs worse than linting in the main thread (i.e. without the `--concurrency` flag or with `--concurrency=off`) because spinning up a worker thread adds extra overhead.

That said, a concurrency value of 1 can be useful for testing or debugging, and it could make sense when using the using the Node.js API, because if other asynchronous tasks are running in the main thread, linting in a separate thread can improve performance. However, the CLI already launches ESLint in its own dedicated process with no competing async work, so specifying `--concurrency=1` there will always degrade performance.

To avoid confusion, we could forbid `--concurrency=1` when ESLint is invoked via the CLI. Another option could be emitting a warning whenever `concurrency` is set to 1, regardless of how ESLint is started. What do you think is the best way to handle this case?</s>

<s>**Should `ESLint.fromOptionModule()` accept Unix pathnames?**
Per the RFC, [`ESLint.fromOptionModule()`](https://github.com/eslint/eslint/blob/rfc129/docs/src/integrate/nodejs-api.md#-eslintfromoptionmoduleoptionsurl) accepts only a URL. I have realized that the original design could be adapted to also accept a Unix pathname like other APIs do. For example, dynamic imports can use absolute paths (`import("/path/to/file.js")`) or relative paths (`import("./file.js")`) besides file URLs. This would be just a minor enhancement to the developer experience, as file paths would continue to be converted to file URLs behind the scenes. Should `ESLint.fromOptionModule()` accept Unix pathnames besides URLs? If so, should relative filenames be allowed and resolved from the current working directory?</s>

<s>**Should TypeScript option modules be loaded like TypeScript config files?**
Thanks to built-in [type stripping](https://nodejs.org/api/typescript.html#type-stripping), Node.js 24 permits loading TypeScript files as option modules out of the box, with a URL like `file:///path/to/eslint-options.ts`. Bun and Deno recognize TypeScript syntax natively. Nonetheless, if we wish to align the way option modules are loaded with how config files are loaded, we could consider allowing native loading of TypeScript files behind a new feature flag, similar to `unstable_native_nodejs_ts_config`, and using jiti otherwise. I'm not sure how useful this approach would be in practice, and I doubt it's worth the added complexity, but it's good to discuss it in advance. Once option modules are released, changing the way files are loaded will be probably a breaking change.</s>

<s>**Should the `concurrency` option be forbidden in the `ESLint` constructor?**
Another point raised in the RFC was the strategy to guarantee that all options are serializable when `concurrency` is enabled. One solution that we did not consider is to completely disallow the `concurrency` option in the `ESLint` constructor for users. This would imply that the only way to specify `concurrency` would be through an option module. While this would eliminate the need to check the options when `concurrency` is specified, it would also sacrifice usability, so I'm not sure if we want to do this.</s>

</details>

<!-- markdownlint-disable-file MD004 -->