@lumirlumir 2025-05-29
Would it make sense to link this issue to https://github.com/eslint/eslint/issues/3565?