@nzakas 2025-06-04
@fasttime it may be helpful to add in some debug statements with profiling data to see where the time is being spent.