@nzakas 2025-06-10
With the latest changes, concurrency=auto runs in roughly the same time as running without concurrency (1m55s), so that's an improvement. 

As an experiment, I eliminated the use of the Retrier completely and that brought the time down to 1m30s. I don't think removing the Retrier completely will work on larger projects, but it at least gives us an idea of where some of the time is being spent.