@nzakas 2025-06-11
I agree, the chances are diminished inside of workers but aren't 0. Because removing the retrier sped things up significantly, it seems worth investigating a lighter-weight option to use inside of workers.