@DreierF 2025-06-12
As you asked for further benchmarks, I also gave the current version a try:
- The codebase has around 2.1k Typescript files
- The test was run on my M1 Max (8 performance, 2 efficiency cores)
- Our config contains a bunch of plugins: typescript-eslint (with typed lining rules), react, react-hooks, storybook

<details><summary>Results</summary>
$ hyperfine --runs 3 --parameter-list threads 1,2,4,6,8,10,auto 'node ./node_modules/.bin/eslint --concurrency {threads}'
Benchmark 1: node ./node_modules/.bin/eslint --concurrency 1
  Time (mean ± σ):     77.193 s ±  0.800 s    [User: 106.509 s, System: 5.332 s]
  Range (min … max):   76.318 s … 77.885 s    3 runs
 
Benchmark 2: node ./node_modules/.bin/eslint --concurrency 2
  Time (mean ± σ):     51.667 s ±  1.110 s    [User: 134.878 s, System: 10.135 s]
  Range (min … max):   50.388 s … 52.381 s    3 runs
 
Benchmark 3: node ./node_modules/.bin/eslint --concurrency 4
  Time (mean ± σ):     41.928 s ±  0.499 s    [User: 184.212 s, System: 24.178 s]
  Range (min … max):   41.579 s … 42.499 s    3 runs
 
Benchmark 4: node ./node_modules/.bin/eslint --concurrency 6
  Time (mean ± σ):     49.671 s ±  3.048 s    [User: 248.316 s, System: 45.340 s]
  Range (min … max):   46.422 s … 52.467 s    3 runs
 
Benchmark 5: node ./node_modules/.bin/eslint --concurrency 8
  Time (mean ± σ):     62.900 s ±  0.948 s    [User: 307.563 s, System: 68.535 s]
  Range (min … max):   61.830 s … 63.636 s    3 runs
 
Benchmark 6: node ./node_modules/.bin/eslint --concurrency 10
  Time (mean ± σ):     86.464 s ±  1.176 s    [User: 397.564 s, System: 99.125 s]
  Range (min … max):   85.337 s … 87.683 s    3 runs
 
Benchmark 7: node ./node_modules/.bin/eslint --concurrency auto
  Time (mean ± σ):     47.486 s ±  2.193 s    [User: 216.097 s, System: 39.932 s]
  Range (min … max):   45.644 s … 49.911 s    3 runs
 
Summary
  node ./node_modules/.bin/eslint --concurrency 4 ran
    1.13 ± 0.05 times faster than node ./node_modules/.bin/eslint --concurrency auto
    1.18 ± 0.07 times faster than node ./node_modules/.bin/eslint --concurrency 6
    1.23 ± 0.03 times faster than node ./node_modules/.bin/eslint --concurrency 2
    1.50 ± 0.03 times faster than node ./node_modules/.bin/eslint --concurrency 8
    1.84 ± 0.03 times faster than node ./node_modules/.bin/eslint --concurrency 1
    2.06 ± 0.04 times faster than node ./node_modules/.bin/eslint --concurrency 10

$ hyperfine --runs 3 'node ./node_modules/.bin/eslint'
Benchmark 1: node ./node_modules/.bin/eslint
  Time (mean ± σ):     60.414 s ±  1.457 s    [User: 88.568 s, System: 4.818 s]
  Range (min … max):   58.772 s … 61.554 s    3 runs
 </details>