@TkDodo 2025-08-13
I made some benchmarks on the [getsentry/sentry](https://github.com/getsentry/sentry) codebase, which has over 1M lines of TypeScript to lint. Our eslint config is [here](https://github.com/getsentry/sentry/blob/435f7e12c51fa4ab8a7191d5c860ce76a524da76/eslint.config.mjs).

Machine stats:

```
Apple M3 Pro
11 cores (5 performance and 6 efficiency)
36GB RAM

node v22.16.0
pnpm v10.10.0
```

before, with eslint 9.32.0:

```
hyperfine --runs=3 'pnpm run lint:js'

Benchmark 1: pnpm run lint:js
  Time (mean ± σ):     202.509 s ±  4.070 s    [User: 348.777 s, System: 23.251 s]
  Range (min … max):   199.176 s … 207.045 s    3 runs
```

with this PR:

```
hyperfine --runs=3 --parameter-list threads off,2,3,4,auto 'pnpm run lint:js --concurrency={threads}'

Benchmark 1: pnpm run lint:js --concurrency=off
  Time (mean ± σ):     189.424 s ±  1.642 s    [User: 330.061 s, System: 20.740 s]
  Range (min … max):   188.384 s … 191.317 s    3 runs
 
Benchmark 2: pnpm run lint:js --concurrency=2
  Time (mean ± σ):     127.073 s ±  2.633 s    [User: 327.366 s, System: 32.953 s]
  Range (min … max):   125.368 s … 130.105 s    3 runs
 
Benchmark 3: pnpm run lint:js --concurrency=3
  Time (mean ± σ):     118.972 s ±  4.122 s    [User: 403.752 s, System: 67.801 s]
  Range (min … max):   116.233 s … 123.714 s    3 runs
 
Benchmark 4: pnpm run lint:js --concurrency=4
  Time (mean ± σ):     131.240 s ± 16.753 s    [User: 477.990 s, System: 118.154 s]
  Range (min … max):   120.795 s … 150.564 s    3 runs
 
Benchmark 5: pnpm run lint:js --concurrency=auto
  Time (mean ± σ):     161.948 s ± 31.493 s    [User: 555.238 s, System: 190.366 s]
  Range (min … max):   140.644 s … 198.123 s    3 runs
```

concurrency 3 performed best, and the savings are substantial - about 35% 🎉

It’s also interesting that from the concurrency setting, `auto` was performing the worst